modified:   machinelearning/regression/LinearRegression.java 	modified:   machinelearning/regressionEx/LinearRegressionEx.java 	modified:   machinelearning/regressionEx/MakeData.java 	modified:   machinelearning/regressionEx/MakeTheta.java 	new file:   machinelearning/regressionEx/TrainData.java 	modified:   MNISTTool/LinearRegressionMNIST.java 	modified:   MNISTTool/MakeTheta.java 	modified:   MNISTTool/TrainData.java 	modified:   ../webapp/WEB-INF/views/home.jsp 	modified:   ../../../target/m2e-wtp/web-resources/META-INF/maven/org.taek/controller/pom.properties

+++ src/main/java/machinelearning/regression/LinearRegression.java
@@ -5,6 +5,6 @@ public interface LinearRegression {
 	// h = X * theta
 	public void hypothesis();
 	// 1/2 * sum(h-y)^2
-	public Long costFuntion();
+	public Double costFuntion();
 	public void gradientDecent();
 }
